+++ src/main/java/me/andreaiacono/racinglearning/rl/RacingMDP.java
@@ -52,7 +52,7 @@ public class RacingMDP implements MDP<ScreenFrameState, Integer, CarActionSpace>
     @Override
     public StepReply<ScreenFrameState> step(Integer action) {
         Command command = CommandsTranslator.getCommandFromInteger(action);
-        long reward = game.move(command);
+        double reward = game.move(command);
         System.out.print("\rEpoch #" + epoch + " - Executed command #" + game.getMovesNumber() + " [" + command + "] - Reward: " + reward + " - Cumulative Reward: " + game.getCumulativeReward() + " - Total moves: " + game.getCumulativeMovesNumber() + "\t\t\t\t\t\t");
         return new StepReply(new ScreenFrameState(screenBuffer), reward, game.isOver(), null);
     }
